Moved BrittleAnnotation to own file
